requirements.txt
Cython~=3.0.10
setuptools~=70.2.0
paho-mqtt~=1.6.1